Now using searchview dialog
