+++ app/src/main/java/com/presentedbykaran/bookshelf/AlertDialogFragment.java
@@ -8,6 +8,8 @@ import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
+import com.presentedbykaran.bookshelf.R;
+
 /**
  * Created by karan on 05/08/18.
  */
