+++ app/src/main/res/menu/options_menu.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <!--<item-->
+    <!--android:id="@+id/app_bar_search"-->
+    <!--android:title="Search"-->
+    <!--app:actionViewClass="android.widget.SearchView"-->
+    <!--app:actionLayout="@xml/searchable"-->
+    <!--app:showAsAction="ifRoom|collapseActionView"-->
+    <!--android:icon="@drawable/ic_search_black_24dp" />-->
+    <item
+        android:id="@+id/searchBtn"
+        android:icon="@drawable/search_white_24dp"
+        android:title="Search"
+        app:showAsAction="always" />
+</menu>
