+++ app/src/main/res/values/strings.xml
@@ -1,7 +1,6 @@
-<resources>
-    <string name="app_name">Bookshelf</string>
+<resources> <string name="app_name">Bookshelf</string>
     <string name="action_settings">Settings</string>
-    <string name="search_hint">Search for an author or book title</string>
+    <string name="search_hint">Search for a book title</string>
     <string name="alert_title">Oops! Sorry</string>
     <string name="error_message">There was an error. Please try again</string>
     <string name="error_button_text">OK</string>
@@ -9,4 +8,5 @@
                                  and try again</string>
     <string name="title_activity_search_results">SearchResultsActivity</string>
     <string name="buttonTextMainScreen">Go</string>
+    <string name="title_activity_main2">Main2Activity</string>
 </resources>
