+++ app/src/main/res/xml/searchable.xml
@@ -2,6 +2,7 @@
 <resources>
     <searchable xmlns:android="http://schemas.android.com/apk/res/android"
         android:label="@string/app_name"
-        android:hint="@string/search_hint">
+        android:hint="@string/search_hint"
+        android:voiceSearchMode="showVoiceSearchButton|launchRecognizer" >
     </searchable>
 </resources>
